correction du temps ca doit etre un double ainsi que le getConstanteK et le robot determine le moment de la collecte

+++ TP1- cheminPlusCourt/src/Robot.cpp
@@ -7,7 +7,7 @@ Robot::Robot()
 }
 
 
-int Robot::getConstanteK() const
+double Robot::getConstanteK() const
 {
 	return constanteK_;
 }
