+++ TP1- cheminPlusCourt/src/Robot.h
@@ -7,7 +7,7 @@ class Robot {
 public:
 	Robot();
 	virtual void setConstanteK(Commande* uneCommande) = 0;
-	int getConstanteK() const;
+	double getConstanteK() const;
 	int getChargeMaximale() const;
 	virtual void afficher() ;
